gradle test failed

+++ bash_deploy/copy.sh
@@ -27,8 +27,11 @@ echo OPTION : ${OPTION}
 
 echo ========== copying ${NAME_PROJECT} file ==========
 
-rm -rf ${FILEPATH}
-echo removed previous ${NAME_PROJECT} file
+if [ -e ${FILEPATH} ]; then
+    rm -rf ${FILEPATH}
+    echo removed previous ${NAME_PROJECT} ${FILEPATH} file
+fi
+
 cp -r ${WORKSPACE}/build/libs/${FILENAME} ${FILEPATH}
 echo copied new ${NAME_PROJECT} file
